requirements.txt
ollama
python-dotenv
huggingface_hub
streamlit
